Claim compatibility with Capybara 2.1.
I don't know how compatible we actually are. Some tests are
failing, but they were failing on 2.0.2 too. Don't have time
to investigate right now, but I need something that says it
works with 2.1 to do some experiments in another project.

In theory 2.1 should work like 2.0, but it has options to
support some of the 1.x functionality we miss. But I'm not
sure kelp should be setting those options.